requirements.txt
langchain
langchain-openai
langchain-community
langchain-docling
faiss-cpu
python-dotenv
transformers
pypdf

# Visualização e análise
matplotlib
scikit-learn
numpy
scipy